osxaudio: Drop in io_proc on pause instead of removing it
Unsetting io_proc while the AudioUnit is rendering audio can deadlock,
so mark the callback as dropping when pausing instead. This is
applicable to both the source and the sink, where we write out
silence when paused.

Another possibility is to stop the AudioUnit when paused, but that is
an intrusive change.

Fixes https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/4155

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/10107>